add ast to parser
